@@ -1,6 +1,6 @@
 {
-  "name": "react-browser-notifications",
-  "version": "1.0.14",
+  "name": "react-browser-notification",
+  "version": "1.0.1",
   "description": "React component for the Javascript Notification API",
   "main": "build/index.js",
   "scripts": {
@@ -10,7 +10,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/knxyzkn/react-browser-notifications.git"
+    "url": "git+https://github.com/Domino987/react-browser-notifications.git"
   },
   "keywords": [
     "react",
@@ -19,12 +19,12 @@
     "api",
     "notifications"
   ],
-  "author": "Kaushik Nagaraj",
+  "author": "Dominik Engel",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/knxyzkn/react-browser-notifications/issues"
+    "url": "https://github.com/Domino987/react-browser-notifications/issues"
   },
-  "homepage": "https://github.com/knxyzkn/react-browser-notifications#readme",
+  "homepage": "https://github.com/Domino987/react-browser-notifications#readme",
   "devDependencies": {
     "babel-cli": "^6.24.1",
     "babel-core": "^6.24.1",
@@ -35,9 +35,9 @@
   },
   "dependencies": {
     "push.js": "^1.0.5",
-    "react": "^16.2.0",
+    "react": "^17.0.0",
     "shortid": "^2.2.8",
-    "webpack": "^2.6.1"
+    "webpack": "^4.0.0"
   },
   "babel": {
     "presets": [
