@@ -1,3 +1,3 @@
-import ReactNotifications from './ReactNotifications.js';
+import ReactNotifications from "./ReactNotifications.js";
 
-export default ReactNotifications;
+export { ReactNotifications };
